Handle promise rejections in the middle of chain
Before, an unhandled promise rejection notice would occur for rejections
that occurred in a chain:

```js
var promise = Promise.resolve().then(() => {
    throw new Error('test');
}).then(() => {
    return 1;
}).catch(() => {
});
```

This happens because the rejected promise never has an `onRejected`
attached to it, and since it is an intermediate value in a promise
chain it never will.

Now, only the end of a promise chain may fire unhandled promise
rejections.